syntax: take the 20% of the retained tree that is padding and 64-bit offsets (#2073) (#2193)
* syntax: stop paying padding and a 64-bit offset on every word (#2073)

A warm interactive startup keeps ~30MB of retained tree, and the profile
says the remaining time is `madvise` and `memclr` — the cost of growing
and zeroing a heap sized by what is *kept*, not by what is thrown away.
Two measurements on this issue already showed that cutting garbage buys
nothing; this cuts what is kept.

Two changes, neither of which moves a semantic:

- `Span`'s fields were declared in reading order — a kind, a value, a
  quoting — and Go padded that out to 88 bytes with 15 bytes of holes.
  Wide fields first and the one-byte ones together at the end makes it
  72. There are ~185,000 spans in the tree a real startup leaves, so the
  holes alone were 3MB of resident memory holding nothing.

- `Pos` was three `int`s, 24 bytes, and it is the most repeated shape in
  a parsed program: a `Word` holds two and every `Span` inside it holds
  another. At `int32` a Pos is 12 bytes, a Word is 48 rather than 72 and
  a Span is 64 rather than 72 — and 48 and 64 are Go size classes, where
  72 was three quarters of one and wasted the rest.

Measured over the 6.3MB of shell the startup path actually reads — 589
files under ~/.zi and ~/.zshrc, parsed and retained:

                   allocated     live at rest
    before           139.2MB          71.9MB
    after            110.5MB          57.5MB
                     -20.6%          -20.1%

The bound `int32` buys that with is 2GiB of input per file, and the
parser already holds the whole of its input as one string, so that is
not the limit anything reaches first.

* syntax: hold the two shapes a tree is made of to a budget (#2073)

The sizes the previous commit bought are the kind that come back
silently. A field added to Span in reading order rather than in width
order gives the padding straight back, nothing else in the repository
notices, and the only symptom is 3MB more resident memory in a shell
somebody started an hour later.

So the numbers are written down with what they cost: one measured
startup keeps ~150,000 Words and ~186,000 Spans, and the test prices a
failure in megabytes rather than reporting a byte count. 48 and 64 are
Go size classes, which is the part that is invisible from the source —
72 bytes is served from the 80-byte class and the spare 8 are paid for
and unreachable.

The second test guards the bound int32 buys those sizes with. A wrapped
offset is a diagnostic pointing at the wrong place, which is silent;
writing 2GiB down as a test makes it a decision rather than an accident
of a type.

Mutation-proven: a string field added to Span fails it at 80 bytes, and
a bool added into existing padding correctly does not.

+++ internal/fmt/comments/comments.go
@@ -54,7 +54,7 @@ func Recover(src string, f *syntax.File) []Comment {
 			end++
 		}
 		out = append(out, Comment{
-			Pos:  syntax.Pos{Offset: i, Line: line, Col: i - lineStart + 1},
+			Pos:  syntax.Pos{Offset: int32(i), Line: int32(line), Col: int32(i - lineStart + 1)},
 			Text: src[i:end],
 		})
 		i = end - 1 // resume at the newline so the line count stays right
@@ -94,18 +94,18 @@ func excluded(f *syntax.File) []span {
 	walk.Nodes(f, func(n syntax.Node) bool {
 		switch x := n.(type) {
 		case *syntax.Word:
-			add(x.Start.Offset, x.Stop.Offset)
+			add(int(x.Start.Offset), int(x.Stop.Offset))
 		case *syntax.Assign:
-			add(x.Start.Offset, x.Stop.Offset)
+			add(int(x.Start.Offset), int(x.Stop.Offset))
 			return false
 		case *syntax.TestClause:
-			add(x.Start.Offset, x.Stop.Offset)
+			add(int(x.Start.Offset), int(x.Stop.Offset))
 			// Fall through to redirections via the walker: they sit
 			// outside [Start, Stop) and carry words of their own.
 		case *syntax.ArithCmdClause:
-			add(x.Start.Offset, x.Stop.Offset)
+			add(int(x.Start.Offset), int(x.Stop.Offset))
 		case *syntax.ForArithClause:
-			add(x.Start.Offset, x.Start.Offset+len(x.Header))
+			add(int(x.Start.Offset), int(x.Start.Offset)+len(x.Header))
 		}
 		return true
 	})
